Struct gmp::rand::RandState

pub struct RandState {
    // some fields omitted
}

Methods

impl RandState

fn new() -> RandState

fn new_mt() -> RandState

fn new_lc_2exp(a: Mpz, c: u64, m2exp: u64) -> RandState

fn new_lc_2exp_size(size: u64) -> RandState

fn seed(&mut self, seed: Mpz)

fn seed_ui(&mut self, seed: u64)

fn urandom(&mut self, n: &Mpz) -> Mpz

Generate a uniform random integer in the range 0 to n-1, inclusive

fn urandom_2exp(&mut self, n: u64) -> Mpz

Generate a uniformly distributed random integer in the range 0 to 2n−1, inclusive.